requirements.txt
streamlit
requests
pandas
scikit-learn
python_dotenv
